requirements.txt
flask==2.3.3
flask-login==0.6.2
Werkzeug==2.3.7
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
gunicorn==20.1.0
pickle-mixin==1.0.2
pillow==9.5.0
cachetools==5.3.1
requests==2.31.0